+++ plugins/imprint/engine/imprint/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.2.1"
+__version__ = "0.2.2"
 
 from .theme import Theme
 from .render import build